Added game state
We need to have game state because when we lose or win the game and if we click on slot again. It will call the function and will go through whole logic even though the game is already finished. Therefore, before continuing with game logic we need to make sure if game is already not finished.